+++ ui/v2.5/src/components/Images/ImageList.tsx
@@ -32,7 +32,7 @@ import { ExportDialog } from "../Shared/ExportDialog";
 import { objectTitle } from "src/core/files";
 import TextUtils from "src/utils/text";
 import { ConfigurationContext } from "src/hooks/Config";
-import { useContainerDimensions } from "../Shared/GridCard";
+import { useContainerDimensions } from "../Shared/GridCard/GridCard";
 
 interface IImageWallProps {
   images: GQL.SlimImageDataFragment[];
